wahoo: Don't override swappiness
I believe there is an incredible misunderstanding when it comes to
how swap is supposed to be used in relation to ZRAM.

Swappiness controls the priority given to pages that are not being
actively used. In doing so, we can conserve memory by compressing
it with ZRAM; the tradeoff being that this adds overhead to the CPU
which in turn must also decompress those pages.

By setting this value to 100, Google effectively made it so that
swapping these *unused* pages to ZRAM is always preferred over
letting them be in memory. This actually isn't all too bad of an
idea on devices with low amounts of RAM, but in the case of wahoo,
we have 4gb ready to be used.

So, essentially, in real world use if a user were to frequently
switch between apps, then we end up with an issue where tasks
are constantly being swapped in and out of ZRAM (which is already
reserved to quite a bit of RAM). When applications being used by
the user are quite big, and switched back and forth to, this can
be rather troubling, and cause things to be slower than they
really should be.

The default of 60 used by the kernel is more than ample for our
needs, and actually handles things quite nicely.

Change-Id: I00019753637f9fb3cfe88b9c1fe0bb22a1193f8a
Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>